requirements.txt
langchain
langgraph
langchain-openai
openai
pandas
numpy
scikit-learn
category-encoders
optuna
xgboost
lightgbm
catboost
matplotlib
python-dotenv
